+++ packages/foam-core/package.json
@@ -15,6 +15,7 @@
     "prepare": "tsdx build"
   },
   "devDependencies": {
+    "@types/dashify": "^1.0.0",
     "@types/graphlib": "^2.1.6",
     "@types/lodash": "^4.14.157",
     "glob": "^7.1.6",
