+++ packages/foam-core/src/utils.ts
@@ -1,11 +1,4 @@
 import { titleCase } from 'title-case';
-import glob from 'glob';
-import { promisify } from 'util';
-import fs from 'fs';
-import { NoteGraph } from './note-graph';
-import { createNoteFromMarkdown } from './markdown-provider';
-
-const findAllFiles = promisify(glob);
 
 export function dropExtension(path: string): string {
   const parts = path.split('.');
@@ -20,23 +13,4 @@ export function dropExtension(path: string): string {
  */
 export const getHeadingFromFileName = (filename: string): string => {
   return titleCase(filename.replace(/[^\w\s]/gi, ' '));
-}
-
-export const initializeNoteGraph = async (aboluteDir: string) => {
-  // remove trailing slash from aboluteDir if exists
-  if (aboluteDir.substr(-1) == '/') aboluteDir = aboluteDir.slice(0, -1);
-
-  const files = await findAllFiles(`${aboluteDir}/**/*.md`, {});
-
-  const graph = new NoteGraph();
-  await Promise.all(
-    (await files).map(f => {
-      return fs.promises.readFile(f).then(data => {
-        const markdown = (data || '').toString();
-        graph.setNote(createNoteFromMarkdown(f, markdown));
-      });
-    })
-  );
-
-  return graph;
 }
