+++ packages/foam-core/types/utils.d.ts
@@ -1,3 +1,1 @@
 declare module 'remark-wiki-link';
-
-declare module 'dashify';
